Improved indentation and arrenged import statement in alphabatically order and made getOrder function's parameters optional (#1wn444v)
