+++ src/store/modules/order/actions.ts
@@ -9,14 +9,14 @@ import { translate } from '@/i18n'
 const actions: ActionTree<OrderState, RootState> = {
   
   // Find Orders
-  async findOrders ( { commit, state, dispatch }, payload) {
+  async findOrders ( { commit, state }, payload) {
     let resp;
     try {
       resp = await OrderService.findOrder(payload)
       if (resp && resp.status === 200 && !hasError(resp)) {
         const orders = resp.data.grouped.orderId;
         // Add some product data into cache to access it later 
-        dispatch('getProductInformation', { orders });
+        this.dispatch('product/getProductInformation', { orders });
 
         if (payload.viewIndex && payload.viewIndex > 0) orders.groups = state.list.items.concat(orders.groups)
         commit(types.ORDER_LIST_UPDATED, {
@@ -29,23 +29,7 @@ const actions: ActionTree<OrderState, RootState> = {
       showToast(translate("Something went wrong"));
     }
     return resp;
-  },
-
-  // Add Product information in cache
-
-  async getProductInformation(context, { orders }){
-    let productIds: any = new Set();
-    orders.groups.forEach((order: any) => {
-      order.doclist.docs.forEach((item: any) =>{
-        if(item.productId) productIds.add(item.productId);
-      })
-    })
-    productIds = [...productIds]
-    if(productIds.length){
-      this.dispatch('product/fetchProducts', { productIds })
-    }
   }
-
 } 
 
 export default actions
