+++ src/store/modules/product/mutations.ts
@@ -8,12 +8,11 @@ const mutations: MutationTree <ProductState> = {
     state.products.total = payload.totalProductsCount;
   },
   [types.PRODUCT_ADD_TO_CACHED_MULTIPLE] (state, payload) {
-      // TODO
-      if (payload.products) {
-          payload.products.forEach((product: any) => {
-              state.cached[product.productId] = product
-          });
-      }
+    if(payload.products){
+      payload.products.forEach((product: any) => {
+        state.cached[product.productId] = product
+      });
+    }
   },
 }
 export default mutations;
